+++ src/net/KNet.Serialization.Avro/KNet.Serialization.Avro.csproj
@@ -18,7 +18,7 @@
 		<None Include="..\..\documentation\articles\usageSerDes.md" Pack="true" PackagePath="\" />
 	</ItemGroup>
 	<ItemGroup>
-	  <PackageReference Include="Apache.Avro" Version="1.11.1" />
+	  <PackageReference Include="Apache.Avro" Version="1.11.3" />
 	  <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1">
 	    <PrivateAssets>all</PrivateAssets>
 	    <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
