+++ src/net/KNet.Serialization.MessagePack/KNet.Serialization.MessagePack.csproj
@@ -21,7 +21,7 @@
 		<PackageReference Include="System.Memory" Version="4.5.5" />
 	</ItemGroup>
 	<ItemGroup>
-	  <PackageReference Include="MessagePack" Version="2.5.108" />
+	  <PackageReference Include="MessagePack" Version="2.5.124" />
 	  <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1">
 	    <PrivateAssets>all</PrivateAssets>
 	    <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
