+++ src/net/KNet.Serialization.Protobuf/KNet.Serialization.Protobuf.csproj
@@ -18,7 +18,7 @@
 		<None Include="..\..\documentation\articles\usageSerDes.md" Pack="true" PackagePath="\" />
 	</ItemGroup>
 	<ItemGroup>
-	  <PackageReference Include="Google.Protobuf" Version="3.23.3" />
+	  <PackageReference Include="Google.Protobuf" Version="3.24.3" />
 	  <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1">
 	    <PrivateAssets>all</PrivateAssets>
 	    <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
